:hammer: refactor: swagger 자물쇠 설정
