+++ Api/src/main/java/hongik/hmut/api/auth/controller/AuthController.java
@@ -77,12 +77,10 @@ public class AuthController {
         return signUpUseCase.registerUserByOICDToken(token, registerRequest);
     }
 
-    @Operation(summary = "개발용 회원가입")
+    @Operation(summary = "개발용 회원가입 및 로그인")
     @Tag(name = "1-2. [카카오]")
     @GetMapping("/oauth/kakao/develop")
-    public AuthResponse registerUserForTest(
-        @RequestParam String code
-    ) {
+    public AuthResponse registerUserForTest(@RequestParam String code) {
         return signUpUseCase.registerUserByKakaoCode(code);
     }
 }
