+++ Api/src/main/java/hongik/hmut/api/auth/model/response/OauthTokenResponse.java
@@ -1,7 +1,7 @@
 package hongik.hmut.api.auth.model.response;
 
 
-import hongik.hmut.outer.api.oauth.dto.KakaoTokenResponse;
+import hongik.hmut.infrastructure.outer.api.oauth.dto.KakaoTokenResponse;
 import lombok.Builder;
 import lombok.Getter;
