+++ Api/src/main/java/hongik/hmut/api/user/service/UserInfoUseCase.java
@@ -1,0 +1,18 @@
+package hongik.hmut.api.user.service;
+
+
+import hongik.hmut.api.user.model.response.UserInfoResponse;
+import hongik.hmut.api.utils.UserUtils;
+import hongik.hmut.core.annotation.UseCase;
+import lombok.RequiredArgsConstructor;
+
+@UseCase
+@RequiredArgsConstructor
+public class UserInfoUseCase {
+
+    private final UserUtils userUtils;
+
+    public UserInfoResponse execute() {
+        return UserInfoResponse.from(userUtils.getUser());
+    }
+}
