+++ Core/src/main/java/hongik/hmut/core/consts/StaticVal.java
@@ -5,7 +5,7 @@ public class StaticVal {
     /*
     토큰 관련 static 변수들
      */
-    public static final String TOKEN_TYPE = "token-type";
+    public static final String TOKEN_TYPE = "type";
     public static final String ACCESS_TOKEN = "ACCESS_TOKEN";
     public static final String REFRESH_TOKEN = "REFRESH_TOKEN";
     public static final String BEARER = "Bearer ";
