+++ Domain/src/main/java/hongik/hmut/domain/domains/user/adaptor/UserAdaptor.java
@@ -4,6 +4,7 @@ package hongik.hmut.domain.domains.user.adaptor;
 import hongik.hmut.core.annotation.Adaptor;
 import hongik.hmut.domain.domains.user.domain.OauthInfo;
 import hongik.hmut.domain.domains.user.domain.User;
+import hongik.hmut.domain.domains.user.exception.UserNotFoundException;
 import hongik.hmut.domain.domains.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 
@@ -19,5 +20,9 @@ public class UserAdaptor {
 
     public User save(User user) {
         return userRepository.save(user);
+    }
+
+    public User query(Long userId) {
+        return userRepository.findById(userId).orElseThrow(() -> UserNotFoundException.EXCEPTION);
     }
 }
