+++ Infrastructure/src/main/java/hongik/hmut/infrastructure/config/KakaoInfoConfig.java
@@ -1,0 +1,24 @@
+package hongik.hmut.infrastructure.config;
+
+
+import feign.codec.Encoder;
+import feign.codec.ErrorDecoder;
+import hongik.hmut.infrastructure.outer.api.oauth.config.KakaoInfoErrorDecoder;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Import;
+
+@Import(KakaoInfoErrorDecoder.class)
+public class KakaoInfoConfig {
+
+    @Bean
+    @ConditionalOnMissingBean(value = ErrorDecoder.class)
+    public KakaoInfoErrorDecoder commonFeignErrorDecoder() {
+        return new KakaoInfoErrorDecoder();
+    }
+
+    @Bean
+    Encoder formEncoder() {
+        return new feign.form.FormEncoder();
+    }
+}
