+++ Infrastructure/src/main/java/hongik/hmut/infrastructure/exception/FeignException.java
@@ -1,0 +1,14 @@
+package hongik.hmut.infrastructure.exception;
+
+
+import hongik.hmut.core.exception.BaseException;
+import hongik.hmut.core.exception.GlobalException;
+
+public class FeignException extends BaseException {
+
+    public static final BaseException EXCEPTION = new FeignException();
+
+    private FeignException() {
+        super(GlobalException.FEIGN_SERVER_ERROR);
+    }
+}
