+++ Infrastructure/src/main/java/hongik/hmut/infrastructure/outer/BaseFeignClientPackage.java
@@ -1,0 +1,3 @@
+package hongik.hmut.infrastructure.outer;
+
+public interface BaseFeignClientPackage {}
