+++ Infrastructure/src/main/java/hongik/hmut/infrastructure/outer/api/oauth/config/KakaoAuthFeignConfig.java
@@ -1,0 +1,21 @@
+package hongik.hmut.infrastructure.outer.api.oauth.config;
+
+
+import feign.codec.Encoder;
+import feign.codec.ErrorDecoder;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
+import org.springframework.context.annotation.Bean;
+
+public class KakaoAuthFeignConfig {
+
+    @Bean
+    @ConditionalOnMissingBean(value = ErrorDecoder.class)
+    public KakaoAuthErrorDecoder commonFeignErrorDecoder() {
+        return new KakaoAuthErrorDecoder();
+    }
+
+    @Bean
+    Encoder formEncoder() {
+        return new feign.form.FormEncoder();
+    }
+}
