+++ Infrastructure/src/main/java/hongik/hmut/infrastructure/outer/api/oauth/dto/OIDCPublicKeyDto.java
@@ -1,0 +1,16 @@
+package hongik.hmut.infrastructure.outer.api.oauth.dto;
+
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+public class OIDCPublicKeyDto {
+
+    private String kid;
+    private String alg;
+    private String use;
+    private String n;
+    private String e;
+}
