+++ Infrastructure/src/main/java/hongik/hmut/infrastructure/outer/api/oauth/dto/OIDCResponse.java
@@ -1,0 +1,13 @@
+package hongik.hmut.infrastructure.outer.api.oauth.dto;
+
+
+import java.util.List;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+public class OIDCResponse {
+
+    List<OIDCPublicKeyDto> oidcPublicKeyDtos;
+}
