--- Infrastructure/src/main/java/hongik/hmut/outer/BaseFeignClientPackage.java
@@ -1,3 +1,0 @@
-package hongik.hmut.outer;
-
-public interface BaseFeignClientPackage {}
