--- Infrastructure/src/main/java/hongik/hmut/outer/api/oauth/config/KakaoInfoClientConfig.java
@@ -1,23 +1,0 @@
-package hongik.hmut.outer.api.oauth.config;
-
-
-import feign.codec.Encoder;
-import feign.codec.ErrorDecoder;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Import;
-
-@Import(KakaoInfoErrorDecoder.class)
-public class KakaoInfoClientConfig {
-
-    @Bean
-    @ConditionalOnMissingBean(value = ErrorDecoder.class)
-    public KakaoInfoErrorDecoder commonFeignErrorDecoder() {
-        return new KakaoInfoErrorDecoder();
-    }
-
-    @Bean
-    Encoder formEncoder() {
-        return new feign.form.FormEncoder();
-    }
-}
