--- Infrastructure/src/main/java/hongik/hmut/outer/api/oauth/dto/KakaoTokenResponse.java
@@ -1,17 +1,0 @@
-package hongik.hmut.outer.api.oauth.dto;
-
-
-import com.fasterxml.jackson.databind.PropertyNamingStrategies.SnakeCaseStrategy;
-import com.fasterxml.jackson.databind.annotation.JsonNaming;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@JsonNaming(SnakeCaseStrategy.class)
-public class KakaoTokenResponse {
-
-    private String accessToken;
-    private String refreshToken;
-    private String idToken;
-}
